Add enabled_release/debug in build_all.sh

+++ build_all.sh
@@ -1,23 +1,29 @@
 clang_version=21
+enabled_release=false
+enabled_debug=true
 
 #pnpm install -g @angular/cli
 
 git submodule update --init --recursive --depth=1
 
 # Build generator with native llvm to prepare wrapper for clang-format
-cmake -S cpp/native -B build_native_release -G Ninja -DCMAKE_BUILD_TYPE=Release || exit 1
-cd build_native_release || exit 1
-pnpm install typescript || exit 1
-export PATH=$PATH:$(pwd)/node_modules/.bin
-cmake --build . --parallel $(nproc --all) || exit 1
-cd .. || exit 1
+if [ "$enabled_release" == "true" ]; then
+    cmake -S cpp/native -B build_native_release -G Ninja -DCMAKE_BUILD_TYPE=Release || exit 1
+    cd build_native_release || exit 1
+    pnpm install typescript || exit 1
+    export PATH=$PATH:$(pwd)/node_modules/.bin
+    cmake --build . --parallel $(nproc --all) || exit 1
+    cd .. || exit 1
+fi
 
-cmake -S cpp/native -B build_native_debug -G Ninja -DCMAKE_BUILD_TYPE=Debug || exit 1
-cd build_native_debug || exit 1
-pnpm install typescript || exit 1
-export PATH=$PATH:$(pwd)/node_modules/.bin
-cmake --build . --parallel $(nproc --all) || exit 1
-cd .. || exit 1
+if [ "$enabled_debug" == "true" ]; then
+    cmake -S cpp/native -B build_native_debug -G Ninja -DCMAKE_BUILD_TYPE=Debug || exit 1
+    cd build_native_debug || exit 1
+    pnpm install typescript || exit 1
+    export PATH=$PATH:$(pwd)/node_modules/.bin
+    cmake --build . --parallel $(nproc --all) || exit 1
+    cd .. || exit 1
+fi
 
 # Build wasm-assembly wrapper
 cd cpp/third_party/emsdk || exit 1
@@ -26,39 +32,59 @@ cd cpp/third_party/emsdk || exit 1
 source ./emsdk_env.sh || exit 1
 cd ../../.. || exit 1
 
-emcmake cmake -S cpp/webassembly -B build_webassembly_release -G "Ninja" -DCMAKE_BUILD_TYPE="Release" || exit 1
-cmake --build build_webassembly_release --parallel $(nproc --all) || exit 1
-mkdir -p angular/src/assets
-rm -f angular/src/assets/*
-cp build_webassembly_release/web* angular/src/assets/ || exit 1
-rm build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.html
-cp build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.* angular/src/assets/ || exit 1
-cp build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.js angular/src/app/features/pdf/services || exit 1
+if [ "$enabled_release" == "true" ]; then
+    emcmake cmake -S cpp/webassembly -B build_webassembly_release -G "Ninja" -DCMAKE_BUILD_TYPE="Release" || exit 1
+    cmake --build build_webassembly_release --parallel $(nproc --all) || exit 1
+    mkdir -p angular/src/assets
+    rm -f angular/src/assets/*
+    cp build_webassembly_release/web* angular/src/assets/ || exit 1
+    rm build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.html
+    cp build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.* angular/src/assets/ || exit 1
+    cp build_webassembly_release/ghostpdl-prefix/src/ghostpdl-build/bin/gs.js angular/src/app/features/pdf/services || exit 1
+fi
 
-emcmake cmake -S cpp/webassembly -B build_webassembly_debug -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON || exit 1
-cmake --build build_webassembly_debug --parallel $(nproc --all) || exit 1
-mkdir -p angular/src/assets
-rm -f angular/src/assets/*
-cp build_webassembly_debug/web* angular/src/assets/ || exit 1
-rm build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.html
-cp build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.* angular/src/assets/ || exit 1
-cp build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.js angular/src/app/features/pdf/services || exit 1
+if [ "$enabled_debug" == "true" ]; then
+    emcmake cmake -S cpp/webassembly -B build_webassembly_debug -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON || exit 1
+    cmake --build build_webassembly_debug --parallel $(nproc --all) || exit 1
+    mkdir -p angular/src/assets
+    rm -f angular/src/assets/*
+    cp build_webassembly_debug/web* angular/src/assets/ || exit 1
+    rm build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.html
+    cp build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.* angular/src/assets/ || exit 1
+    cp build_webassembly_debug/ghostpdl-prefix/src/ghostpdl-build/bin/gs.js angular/src/app/features/pdf/services || exit 1
+fi
 
 if [ -f /usr/lib/llvm/${clang_version}/bin/clang ]; then
-    cmake -S cpp/tests/ -B build_tests_debug -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DCMAKE_C_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang -DCMAKE_CXX_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang++ -DCMAKE_AR=/usr/lib/llvm/${clang_version}/bin/llvm-ar -DCMAKE_AS=/usr/lib/llvm/${clang_version}/bin/llvm-as -DCMAKE_RANLIB=/usr/lib/llvm/${clang_version}/bin/llvm-ranlib -DCMAKE_LINKER_TYPE=LLD || exit 1
-    cmake -S cpp/tests/ -B build_tests_release -DWITH_SANITIZE_ADDRESS=OFF -DWITH_SANITIZE_UNDEFINED=OFF -G "Ninja" -DCMAKE_BUILD_TYPE="Release" -DCMAKE_C_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang -DCMAKE_CXX_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang++ -DCMAKE_AR=/usr/lib/llvm/${clang_version}/bin/llvm-ar -DCMAKE_AS=/usr/lib/llvm/${clang_version}/bin/llvm-as -DCMAKE_RANLIB=/usr/lib/llvm/${clang_version}/bin/llvm-ranlib -DCMAKE_LINKER_TYPE=LLD || exit 1
+    if [ "$enabled_debug" == "true" ]; then
+        cmake -S cpp/tests/ -B build_tests_debug -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DCMAKE_C_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang -DCMAKE_CXX_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang++ -DCMAKE_AR=/usr/lib/llvm/${clang_version}/bin/llvm-ar -DCMAKE_AS=/usr/lib/llvm/${clang_version}/bin/llvm-as -DCMAKE_RANLIB=/usr/lib/llvm/${clang_version}/bin/llvm-ranlib -DCMAKE_LINKER_TYPE=LLD || exit 1
+    fi
+    if [ "$enabled_release" == "true" ]; then
+        cmake -S cpp/tests/ -B build_tests_release -DWITH_SANITIZE_ADDRESS=OFF -DWITH_SANITIZE_UNDEFINED=OFF -G "Ninja" -DCMAKE_BUILD_TYPE="Release" -DCMAKE_C_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang -DCMAKE_CXX_COMPILER=/usr/lib/llvm/${clang_version}/bin/clang++ -DCMAKE_AR=/usr/lib/llvm/${clang_version}/bin/llvm-ar -DCMAKE_AS=/usr/lib/llvm/${clang_version}/bin/llvm-as -DCMAKE_RANLIB=/usr/lib/llvm/${clang_version}/bin/llvm-ranlib -DCMAKE_LINKER_TYPE=LLD || exit 1
+    fi
 else
-    cmake -S cpp/tests/ -B build_tests_debug -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DCMAKE_C_COMPILER=/usr/bin/clang-${clang_version} -DCMAKE_CXX_COMPILER=/usr/bin/clang++-${clang_version} -DCMAKE_AR=/usr/bin/llvm-ar-${clang_version} -DCMAKE_AS=/usr/bin/llvm-as-${clang_version} -DCMAKE_RANLIB=/usr/bin/llvm-ranlib-${clang_version} -DCMAKE_LINKER_TYPE=LLD || exit 1
-    cmake -S cpp/tests/ -B build_tests_release -DWITH_SANITIZE_ADDRESS=OFF -DWITH_SANITIZE_UNDEFINED=OFF -G "Ninja" -DCMAKE_BUILD_TYPE="Release" -DCMAKE_C_COMPILER=/usr/bin/clang-${clang_version} -DCMAKE_CXX_COMPILER=/usr/bin/clang++-${clang_version} -DCMAKE_AR=/usr/bin/llvm-ar-${clang_version} -DCMAKE_AS=/usr/bin/llvm-as-${clang_version} -DCMAKE_RANLIB=/usr/bin/llvm-ranlib-${clang_version} -DCMAKE_LINKER_TYPE=LLD || exit 1
+    if [ "$enabled_debug" == "true" ]; then
+        cmake -S cpp/tests/ -B build_tests_debug -DWITH_SANITIZE_ADDRESS=ON -DWITH_SANITIZE_UNDEFINED=ON -G "Ninja" -DCMAKE_BUILD_TYPE="Debug" -DCMAKE_C_COMPILER=/usr/bin/clang-${clang_version} -DCMAKE_CXX_COMPILER=/usr/bin/clang++-${clang_version} -DCMAKE_AR=/usr/bin/llvm-ar-${clang_version} -DCMAKE_AS=/usr/bin/llvm-as-${clang_version} -DCMAKE_RANLIB=/usr/bin/llvm-ranlib-${clang_version} -DCMAKE_LINKER_TYPE=LLD || exit 1
+    fi
+    if [ "$enabled_release" == "true" ]; then
+        cmake -S cpp/tests/ -B build_tests_release -DWITH_SANITIZE_ADDRESS=OFF -DWITH_SANITIZE_UNDEFINED=OFF -G "Ninja" -DCMAKE_BUILD_TYPE="Release" -DCMAKE_C_COMPILER=/usr/bin/clang-${clang_version} -DCMAKE_CXX_COMPILER=/usr/bin/clang++-${clang_version} -DCMAKE_AR=/usr/bin/llvm-ar-${clang_version} -DCMAKE_AS=/usr/bin/llvm-as-${clang_version} -DCMAKE_RANLIB=/usr/bin/llvm-ranlib-${clang_version} -DCMAKE_LINKER_TYPE=LLD || exit 1
+    fi
 fi
-cmake --build build_tests_debug --parallel $(nproc --all) || exit 1
-cmake --build build_tests_release --parallel $(nproc --all) || exit 1
-cd build_tests_debug || exit 1
-ctest || exit 1
-cd .. || exit 1
-cd build_tests_release || exit 1
-ctest || exit 1
-cd .. || exit 1
+if [ "$enabled_debug" == "true" ]; then
+    cmake --build build_tests_debug --parallel $(nproc --all) || exit 1
+fi
+if [ "$enabled_release" == "true" ]; then
+    cmake --build build_tests_release --parallel $(nproc --all) || exit 1
+fi
+if [ "$enabled_debug" == "true" ]; then
+    cd build_tests_debug || exit 1
+    ctest || exit 1
+    cd .. || exit 1
+fi
+if [ "$enabled_release" == "true" ]; then
+    cd build_tests_release || exit 1
+    ctest || exit 1
+    cd .. || exit 1
+fi
 
 # Build Angular project
 cd angular || exit 1
